market_data_service/pyproject.toml
[project]
name = "market-data-service"
version = "0.1.0"
description = "A Model Context Protocol server providing market data"
requires-python = ">=3.10"
dependencies = [
 "aiohttp>=3.11.8",
 "ib-insync>=0.9.86",
 "mcp[cli]>=1.0.0",
 "mplfinance>=0.12.10b0",
 "nest-asyncio>=1.6.0",
 "pandas-ta>=0.3.14b0",
 "pandas>=2.2.3",
 "pydantic>=2.10.2",
 "python-dotenv>=1.0.1",
 "requests>=2.32.3",
 "retry>=0.9.2",
 "numpy<2.0.0",
 "setuptools>=75.6.0",
]

[[project.authors]]
name = "Itay Verkh"
email = "itay1542@gmail.com"


[project.scripts]
market_data_server = "src.server:main"
